Adjust DLL support + Env layer
