+++ Runtime/AuthLayer/EOSAuth.cs
@@ -290,86 +290,5 @@ namespace RobProductions.OpenEOS
 			return LoginConnect(platformSession, connectSet, loginCompleteCallback);
 		}
 
-		//COMMAND LINE HELPER
-
-		/// <summary>
-		/// As per https://dev.epicgames.com/docs/epic-account-services/auth/auth-interface,
-		/// when launched through the Epic Games Launcher,
-		/// the app's command line will be given 3 special codes:
-		/// AUTH_LOGIN, AUTH_TYPE, and AUTH_PASSWORD,
-		/// of which one is useful (password, which acts as the credential token)
-		/// but I will check Auth Type just in case. Use this token
-		/// to easily log in a user (to Auth service for EpicAccountId, using the ExchangeCode mode)
-		/// without requiring the web browser.
-		/// </summary>
-		/// <returns>The token parsed from the command line</returns>
-		public static string GetExchangeCodeToken()
-		{
-			var authTypeCode = GetCommandLineArgValue("-AUTH_TYPE");
-			if(authTypeCode != null && authTypeCode == "exchangecode")
-			{
-				var pass = GetCommandLineArgValue("-AUTH_PASSWORD");
-				if(pass == null)
-				{
-					EOSCore.LogEOS("Warning: AUTH_PASSWORD was null!");
-				}
-
-				return pass;
-			}
-			else
-			{
-				EOSCore.LogEOS("Warning: AUTH_TYPE was not exchangecode");
-			}
-
-			return null;
-		}
-
-		/// <summary>
-		/// Split a full command argument into its value.
-		/// </summary>
-		/// <param name="argName"></param>
-		/// <returns>The value after the = sign in a command line argument</returns>
-		public static string GetCommandLineArgValue(string argName)
-		{
-			var fullVal = GetCommandLineArg(argName);
-			if(fullVal != null)
-			{
-				var valSplit = fullVal.Split(new[] { '=' }, 2);
-				if(valSplit.Length > 1)
-				{
-					return valSplit[1];
-				}
-				else
-				{
-					EOSCore.LogEOS("Warning: CommandLineArg " + fullVal + " could not be split with seperator");
-				}
-			}
-			else
-			{
-				EOSCore.LogEOS("Warning: CommandLineArg " + argName + " did not exist in System environment");
-			}
-			return null;
-		}
-
-		/// <summary>
-		/// Returns an argument name that starts with the given string.
-		/// Epic Launcher formats some arguments like -NAME=VALUE,
-		/// so we want to detect the given name and return the full string
-		/// to parse out the value.
-		/// </summary>
-		/// <param name="startsWith"></param>
-		/// <returns>The full string of the argument</returns>
-		public static string GetCommandLineArg(string startsWith)
-		{
-			var args = System.Environment.GetCommandLineArgs();
-			for (int i = 0; i < args.Length; i++)
-			{
-				if (args[i].StartsWith(startsWith))
-				{
-					return args[i];
-				}
-			}
-			return null;
-		}
 	}
 }
