--- Runtime/EOSSDK/SDK/Plugins/Android.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1f4dd81c636bfab4faf7c9d233f13dc8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
