--- Runtime/EOSSDK/SDK/Plugins/Android/dynamic-stdc++.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 26e661868590e4e40a875a47b9b681c9
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
