--- Runtime/EOSSDK/SDK/Plugins/Android/dynamic-stdc++/aar.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 40013a73e59a5694ea79237be7651aa8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
