--- Runtime/EOSSDK/SDK/Plugins/Android/dynamic-stdc++/libs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 063cd1906b428fe44a0747d8c4b7ae11
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
