--- Runtime/EOSSDK/SDK/Plugins/Android/dynamic-stdc++/libs/arm64-v8a.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ee8a562234bfaf64ab9f0af1d11508b8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
