--- Runtime/EOSSDK/SDK/Plugins/Android/dynamic-stdc++/libs/arm64-v8a/libEOSSDK.so.meta
@@ -1,27 +1,0 @@
-fileFormatVersion: 2
-guid: 9a82058906ea13144904bf9831d23ce7
-PluginImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  iconMap: {}
-  executionOrder: {}
-  defineConstraints: []
-  isPreloaded: 0
-  isOverridable: 0
-  isExplicitlyReferenced: 0
-  validateReferences: 1
-  platformData:
-  - first:
-      Any: 
-    second:
-      enabled: 1
-      settings: {}
-  - first:
-      Editor: Editor
-    second:
-      enabled: 0
-      settings:
-        DefaultValueInitialized: true
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
