--- Runtime/EOSSDK/SDK/Plugins/Android/dynamic-stdc++/libs/armeabi-v7a.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fd4263018cea8634b8d49e5633ff4161
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
