--- Runtime/EOSSDK/SDK/Plugins/Android/static-stdc++.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fefc18e43c9d3c7469ec170b8d99ea31
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
