--- Runtime/EOSSDK/SDK/Plugins/Android/static-stdc++/aar.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 69507f159e8d5e04d9c0a4d9b93681ec
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
