--- Runtime/EOSSDK/SDK/Plugins/Android/static-stdc++/libs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 98765c9715402d64281787632b423c06
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
