--- Runtime/EOSSDK/SDK/Plugins/Android/static-stdc++/libs/arm64-v8a.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0ad7b38d86adde14bae1b0ea38683cf4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
