--- Runtime/EOSSDK/SDK/Plugins/Android/static-stdc++/libs/armeabi-v7a.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 842b36dbe974b9b4fa6bcc004f52d30c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
