--- Runtime/EOSSDK/SDK/Plugins/Android/static-stdc++/libs/armeabi-v7a/libEOSSDK.so.meta
@@ -1,27 +1,0 @@
-fileFormatVersion: 2
-guid: ba557ce8bc1f982479dff30eb36f29a3
-PluginImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  iconMap: {}
-  executionOrder: {}
-  defineConstraints: []
-  isPreloaded: 0
-  isOverridable: 0
-  isExplicitlyReferenced: 0
-  validateReferences: 1
-  platformData:
-  - first:
-      Any: 
-    second:
-      enabled: 1
-      settings: {}
-  - first:
-      Editor: Editor
-    second:
-      enabled: 0
-      settings:
-        DefaultValueInitialized: true
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
