--- Runtime/EOSSDK/SDK/Plugins/IOS.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: bd63ec7e9be9d6645ad2e8cf6bf992aa
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
