--- Runtime/EOSSDK/SDK/Plugins/IOS/EOSSDK.framework/EOSSDK.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: a4674a493a2efda49b740800c216ade6
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
