--- Runtime/EOSSDK/SDK/Plugins/IOS/EOSSDK.framework/Info.plist
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>BuildMachineOSBuild</key>
-	<string>16E195</string>
-	<key>CFBundleDevelopmentRegion</key>
-	<string>en</string>
-	<key>CFBundleExecutable</key>
-	<string>EOSSDK</string>
-	<key>CFBundleIdentifier</key>
-	<string>com.epicgames.EOSSDK.EOSSDK.Embedded</string>
-	<key>CFBundleInfoDictionaryVersion</key>
-	<string>6.0</string>
-	<key>CFBundleName</key>
-	<string>EOSSDK</string>
-	<key>CFBundlePackageType</key>
-	<string>FMWK</string>
-	<key>CFBundleShortVersionString</key>
-	<string>1.0</string>
-	<key>CFBundleSupportedPlatforms</key>
-	<array>
-		<string>iPhoneOS</string>
-	</array>
-	<key>CFBundleVersion</key>
-	<string>1</string>
-	<key>DTCompiler</key>
-	<string>com.apple.compilers.llvm.clang.1_0</string>
-	<key>DTPlatformBuild</key>
-	<string>15A327</string>
-	<key>DTPlatformName</key>
-	<string>iphoneos</string>
-	<key>DTPlatformVersion</key>
-	<string>11.0</string>
-	<key>DTSDKBuild</key>
-	<string>15A327</string>
-	<key>DTSDKName</key>
-	<string>iphoneos11.0.internal</string>
-	<key>DTXcode</key>
-	<string>0900</string>
-	<key>DTXcodeBuild</key>
-	<string>9M174d</string>
-	<key>MinimumOSVersion</key>
-	<string>11.0</string>
-	<key>UIDeviceFamily</key>
-	<array>
-		<integer>1</integer>
-		<integer>2</integer>
-	</array>
-	<key>UIRequiredDeviceCapabilities</key>
-	<array>
-		<string>arm64</string>
-	</array>
-</dict>
-</plist>
