--- Runtime/EOSSDK/SDK/Plugins/IOS/EOSSDK.framework/Info.plist.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9f4097602e18c5748bac6022e13c2f5a
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
