+++ Runtime/EnvLayer/EOSEnv.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 55c494e1874523341ae92f9629a77e82
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
